+++ src/main/java/quiztastic/core/Player.java
@@ -1,0 +1,31 @@
+package quiztastic.core;
+
+import java.util.Objects;
+
+public class Player {
+    private final String id;
+
+    public Player(String id) {
+        this.id = id;
+    }
+
+    @Override
+    public String toString() {
+        return "Player{" +
+                "id='" + id + '\'' +
+                '}';
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+        Player player = (Player) o;
+        return Objects.equals(id, player.id);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(id);
+    }
+}
